Renamed WebSocket namespace to WebSockets
