+++ Cryptodd/Binance/Orderbooks/Websockets/BaseBinanceOrderbookWebsocketOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace Cryptodd.Binance.Orderbooks.Websocket;
+﻿namespace Cryptodd.Binance.Orderbooks.Websockets;
 
 public abstract class BaseBinanceOrderbookWebsocketOptions
 {
